refactored greeting
